Adding Lua support as a full runtime and framework using Luajit beta7
Lua support via wsapi handlers. Any file named *.lua or *.ws will be
evaluated as a lua cgi script via wsapi and its assorted apis.
very similar to rack in ruby

+++ cloud_controller/spec/support/staging_spec_helpers.rb
@@ -59,7 +59,7 @@ module StagingSpecHelpers
     nil
   ensure
     if block_given?
-      FileUtils.rm_r(working_dir)
+      FileUtils.rm_r(working_dir) if working_dir
     end
     FileUtils.rm_r(source_tempdir) if source_tempdir
   end
